Refactored database code and custom middlewares
