+++ .env.example
@@ -6,7 +6,7 @@
 MONGODB_URI=mongodb://<mlab_user>:<mlab_password>@<mlab_connection_url>
 
 # This is standard running mongodb locally
-MONGODB_URI_LOCAL=mongodb://localhost:27017
+MONGODB_URI_LOCAL=mongodb://admin:adminadmin@localhost:27017
 
 # Put lots of randomness in these
 SESSION_SECRET=ashdfjhasdlkjfhalksdjhflak
