+++ src/app.ts
@@ -7,13 +7,11 @@ import errorHandler from "errorhandler";
 import logger from "./util/logger";
 import lusca from "lusca";
 import dotenv from "dotenv";
-import mongo from "connect-mongo";
 import flash from "express-flash";
 import path from "path";
 import mongoose from "mongoose";
 import passport from "passport";
 import expressValidator from "express-validator";
-import bluebird from "bluebird";
 import { MONGODB_URI, SESSION_SECRET } from "./util/secrets";
 
 // Controllers (route handlers)
@@ -21,6 +19,8 @@ import { HomeController, API, ContactController, UserController, OAuth } from ".
 
 import { Provider } from "./core/provider";
 import { app } from "./core/app";
+import { DBSetup, store } from "./db/setup";
+import { LoginGuard } from "./middlewares";
 
 @app({
   controllers: [HomeController, API, ContactController, UserController, OAuth], // Add new controllers here to register
@@ -32,23 +32,10 @@ import { app } from "./core/app";
     },
     middlewares: [
       errorHandler(), compression(), bodyParser.json(), bodyParser.urlencoded({ extended: true }), expressValidator(),
-      session({ resave: true, saveUninitialized: true, secret: SESSION_SECRET }),
+      session({ resave: true, saveUninitialized: true, secret: SESSION_SECRET, store: store }),
       passport.initialize(), passport.session(), flash(), lusca.xframe("SAMEORIGIN"), lusca.xssProtection(true),
       (req: Request, res: Response, next: any) => { res.locals.user = req.user; next(); },
-      (req: Request, res: Response, next: any) => {
-        // After successful login, redirect back to the intended page
-        if (!req.user &&
-          req.path !== "/login" &&
-          req.path !== "/signup" &&
-          !req.path.match(/^\/auth/) &&
-          !req.path.match(/\./)) {
-          req.session.returnTo = req.path;
-        } else if (req.user &&
-          req.path == "/account") {
-          req.session.returnTo = req.path;
-        }
-        next();
-      },
+      new LoginGuard().index(),
       express.static(path.join(__dirname, "public"), { maxAge: 31557600000 })
     ]
   }
@@ -57,13 +44,12 @@ export class App {
   private _express: Express;
   private env: string;
   private appState: string;
+  private static _app: App = undefined;
 
+  private constructor() { }
   get express(): Express { return this._express; }
   set express(value: Express) { this._express = value; }
 
-  private constructor() { }
-
-  private static _app: App = undefined;
   static getApp(env: string): App {
     if (!this._app) {
       this._app = new App();
@@ -75,32 +61,16 @@ export class App {
   execute = async () => {
     if (this.appState == "executed") return;
 
-    const MongoStore = mongo(session);
-
     // Load environment variables from .env file, where API keys and passwords are configured
     dotenv.config({ path: ".env.example" });
-
-    // Connect to MongoDB
-    const mongoUrl = MONGODB_URI;
-    (<any>mongoose).Promise = bluebird;
-    mongoose.connect(mongoUrl, {useMongoClient: true}).then(
-      () => { /** ready to use. The `mongoose.connect()` promise resolves to undefined. */ },
-    ).catch(err => {
-      console.log("MongoDB connection error. Please make sure MongoDB is running. " + err);
-      // process.exit();
-    });
-
+    DBSetup.Initialize();
     this.appState = "executed";
   }
 
-  async listen(port: number) {
+  listen = async (port: number) => {
     const promise = await new Promise((resolve, reject) => {
       this.express.listen(port, () => {
-        console.log(
-          "  App is running at http://localhost:%d in %s mode",
-          port,
-          this.env
-        );
+        console.log("  App is running at http://localhost:%d in %s mode", port, this.env );
         console.log("  Press CTRL-C to stop\n");
         resolve();
       });
